TP-1100: Removing raml-maven-test-utils module (which was moved in to it's own repo) and some general tidy up
