+++ raml-maven-plugin/src/test/java/uk/gov/justice/raml/maven/generator/GenerateMojoTest.java
@@ -8,8 +8,8 @@ import static org.hamcrest.Matchers.hasProperty;
 import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.core.IsEqual.equalTo;
 
+import uk.gov.justice.maven.test.utils.BetterAbstractMojoTestCase;
 import uk.gov.justice.raml.core.GeneratorConfig;
-import uk.gov.justice.raml.maven.test.utils.BetterAbstractMojoTestCase;
 
 import java.io.File;
 import java.nio.file.Files;
