+++ raml-maven-plugin/src/test/java/uk/gov/justice/raml/maven/validator/RamlSyntaxCheckMojoTest.java
@@ -9,8 +9,8 @@ import static org.mockito.Matchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import uk.gov.justice.maven.test.utils.BetterAbstractMojoTestCase;
 import uk.gov.justice.raml.maven.common.BasicGoalConfig;
-import uk.gov.justice.raml.maven.test.utils.BetterAbstractMojoTestCase;
 
 import java.io.File;
 import java.io.IOException;
